requirements.txt
streamlit
Pillow
groq